Make sure that there's no cert data available if want to *enforce* usage of self-signed certificates.
Take Cheuk's fix and add renamed flag, which reflects the actual intent.
The k8s code woesn't allow usage of self-signed certificates in case a certificate is provided.

Some providers supply certificates with a proper CA chain, some supply self-signed certificates.
So, the flag is really "enforce usage of self-signed certificates", in which case both CAFile and CAData
have to be cleared out.