Add the flag to signify whether or not self-signed certificates use will be enforced.
We use two flags: One hardcoded logic which identifies whether we are working with the OpenShift cluster.
Another is the flag passed in through the command line.
If we are working with the OpenShift cluster, which means we are in the environment where we must use proper certificates, then we don't force self-signed certs.
Otherwise, then we use whatever flag we passed through the command line.