Force deployment to restart pods when configmap changes Using this method: https://helm.sh/docs/howto/charts_tips_and_tricks/#automatically-roll-deployments
When the configmap changes, this will cause the checksum to change,
causing the annotation to change.

This will cause the pod spec to change, which will mean that the
Deployment creates a new pod.

This PR is a remake of https://github.com/turbonomic/kubeturbo/pull/798,
since it was closed without reason.